refactor: extract ErrorGroupList from TabErrors
